test for `@cufunc` replace `^` & manually convert to Int32 in `literal_pow`

+++ src/broadcast.jl
@@ -38,7 +38,7 @@ for f in libdevice
 end
 
 #broadcast ^
-culiteral_pow(::typeof(^), x::Union{Float32, Float64}, ::Val{p}) where p = CUDAnative.pow(x, p)
+culiteral_pow(::typeof(^), x::Union{Float32, Float64}, ::Val{p}) where p = CUDAnative.pow(x, Int32(p))
 
 cufunc(::typeof(Base.literal_pow)) = culiteral_pow
 cufunc(::typeof(Base.:(^))) = CUDAnative.pow
